+++ tests/unit_tests/extrinsics/test_prometheus.py
@@ -1,7 +1,7 @@
 import pytest
 from unittest.mock import MagicMock, patch
 import bittensor
-from bittensor.subtensor import subtensor as Subtensor
+from bittensor import subtensor as Subtensor
 from bittensor.wallet import wallet as Wallet
 from bittensor.extrinsics.prometheus import prometheus_extrinsic
